@@ -1,6 +1,6 @@
 module go-liteflow
 
-go 1.18
+go 1.22
 
 require (
 	google.golang.org/grpc v1.33.1
